Implement getBookAuthors and getBookGenres

+++ 05-spring-jdbc/src/main/java/ru/shishmakov/dao/AuthorDao.java
@@ -30,7 +30,7 @@ public class AuthorDao implements Dao<Author> {
     }
 
     @Override
-    public Author getById(int id) {
+    public Author getById(long id) {
         return null;
     }
