+++ 05-spring-jdbc/src/main/java/ru/shishmakov/dao/GenreDao.java
@@ -30,7 +30,7 @@ public class GenreDao implements Dao<Genre> {
     }
 
     @Override
-    public Genre getById(int id) {
+    public Genre getById(long id) {
         return null;
     }
